# packages/portal-contracts/openapi/portal-api-v1.json
{
  "components": {
    "schemas": {
...
      "CsrfView": {
        "additionalProperties": false,
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          }
        },
        "required": [
          "csrf_token",
          "generation"
        ],
        "title": "CsrfView",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Portal-wide error response contract.",
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "dependency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "documentation_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentation Url"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "error_code",
          "correlation_id",
          "timestamp"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "SessionView": {
        "additionalProperties": false,
        "properties": {
          "absolute_expires_at": {
            "format": "date-time",
            "title": "Absolute Expires At",
            "type": "string"
          },
          "assurance": {
            "title": "Assurance",
            "type": "string"
          },
          "authenticated_at": {
            "format": "date-time",
            "title": "Authenticated At",
            "type": "string"
          },
          "capability_revision": {
            "title": "Capability Revision",
            "type": "string"
          },
          "environment_ids": {
            "items": {
              "type": "string"
            },
            "title": "Environment Ids",
            "type": "array"
          },
          "idle_expires_at": {
            "format": "date-time",
            "title": "Idle Expires At",
            "type": "string"
          },
          "policy_revision": {
            "title": "Policy Revision",
            "type": "string"
          },
          "principal_reference": {
            "format": "uuid",
            "title": "Principal Reference",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "session_reference": {
            "format": "uuid",
            "title": "Session Reference",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "session_reference",
          "principal_reference",
          "tenant_id",
          "status",
          "roles",
          "environment_ids",
          "assurance",
          "policy_revision",
          "capability_revision",
          "authenticated_at",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "SessionView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Data Platform Engineering"
    },
    "description": "PR-PORTAL-001 exposes safe foundation health and system metadata only. No data-platform business capabilities or mutations are enabled.",
    "license": {
      "name": "Private repository"
    },
    "summary": "Foundation BFF for the Enterprise Data Platform Portal",
    "title": "Fintech Data Platform Portal API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/session": {
      "get": {
        "operationId": "getSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "session"
        ]
      }
    },
    "/v1/session/csrf": {
      "get": {
        "operationId": "getSessionCsrf",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "Get Session Csrf",
        "tags": [
          "session"
        ]
      }
    },
...
    "/v1/session/refresh": {
      "post": {
        "operationId": "refreshSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "Refresh Session",
        "tags": [
          "session"
        ]
      }
    },
...
  }
}